+++ changes.md
@@ -1,8 +1,9 @@
 # Differences to the WordPress Repository version
 
-* new shortcode: overviewmap
+* new shortcode overviewmap: generates an overview map with geo positions provided in the pages and posts
 * new options for hover: class (style the tooltip) and popupclose (keep the popup open or not)
 * reduce inline Javascript for geojsonmarker
+* multielevation accepts now also kml files (may work)
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).